requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.128.0
python-dotenv>=1.0.0
pydantic>=2.7.0
apscheduler>=3.10.4
